[Favorites] Rename queries to reflect they are about faves.
